Rdzv manager to check NCCL
