+++ dlrover/python/master/node/event_callback.py
@@ -21,7 +21,7 @@ from dlrover.python.common.constants import (
 )
 from dlrover.python.common.log import default_logger as logger
 from dlrover.python.master.elastic_training.rdzv_manager import (
-    RendezvousManager,
+    ElasticTrainingRendezvousManager,
 )
 from dlrover.python.master.monitor.speed_monitor import SpeedMonitor
 from dlrover.python.master.watcher.base_watcher import Node
@@ -210,7 +210,9 @@ class AllReduceNodeHandlingCallback(NodeEventCallback):
         super(AllReduceNodeHandlingCallback, self).__init__()
         self._master = master
         self._speed_monitor: SpeedMonitor = self._master.speed_monitor
-        self._rdzv_manager: RendezvousManager = self._master.rdzv_manager
+        self._rdzv_manager: ElasticTrainingRendezvousManager = (
+            self._master.rdzv_manager
+        )
 
     def get_job_exit_reason(self, node: Node):
         if self._master.task_manager.training_started():
